Renamed SubCommand sample to Subcommand
