+++ src/Ookii.CommandLine.sln
@@ -17,7 +17,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ookii.CommandLine.Tests", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Parser", "Samples\Parser\Parser.csproj", "{A8F86838-1956-4046-A8D9-93EBC301A468}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SubCommand", "Samples\SubCommand\SubCommand.csproj", "{329D012F-AD31-4D50-AE3C-8F500CDF62FA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Subcommand", "Samples\Subcommand\Subcommand.csproj", "{329D012F-AD31-4D50-AE3C-8F500CDF62FA}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{DC9CCD22-9B9B-4298-8C68-BC7A5A680F93}"
 EndProject
