--- src/Samples/SubCommand/ExitCode.cs
@@ -1,10 +1,0 @@
-﻿namespace SubCommand
-{
-    internal enum ExitCode
-    {
-        Success,
-        CreateCommandFailure,
-        ReadWriteFailure,
-        FileExists,
-    }
-}
