+++ src/Samples/Subcommand/ExitCode.cs
@@ -1,0 +1,8 @@
+﻿namespace SubcommandSample;
+internal enum ExitCode
+{
+    Success,
+    CreateCommandFailure,
+    ReadWriteFailure,
+    FileExists,
+}
